+++ packages/gamut-labs/src/index.ts
@@ -14,8 +14,6 @@ export * from './Banner';
 export * from './BetaSticker';
 export * from './Byline';
 export * from './CurriculumCard';
-export * from './CurriculumCard/Difficulty';
-export * from './CurriculumCard/Difficulty/types';
 export * from './DropdownButton';
 export * from './EditorialImage';
 export * from './EmptySection';
